==> LOGSverification&Processing&Variables/2_MLP_Cleaning.ipynb <==
#basic package
import tqdm
import os
import numpy as np
import glob
import pandas as pd
import sys
import time
import datetime as dt
import itertools
import re
import math
from collections import Counter
import operator

#scaling
from sklearn.preprocessing import StandardScaler, MinMaxScaler

#modelling
from catboost import CatBoostClassifier, Pool

#plot
import matplotlib.pyplot as plt
import seaborn as sns

PACKAGE_PARENT = '../'
SCRIPT_DIR = os.path.dirname(os.path.realpath(os.path.join(os.getcwd(), os.path.expanduser('__file__'))))
sys.path.append(os.path.normpath(os.path.join(SCRIPT_DIR, PACKAGE_PARENT)))
from UTILS import preprocessing_Origins, is_day, is_WG_open, openDevice, cleaning_processing, FB_process, FB_daily
import config_origins as config

# Define parameters

print('change the configuration file if not done yet!')
path_extracted_data = config.path_extracted_data
path_initial_data = config.path_initial_data
id_run = config.id_run
dico_matching = config.dico_matching
nbr_maxdur2beremoved = config.nbr_maxdur2beremoved
dico_garden_opening_hour = config.dico_garden_opening_hour
date_first_opening_WG = config.date_first_opening_WG
print(id_run)
path_extracted_data_visual = os.path.join(path_extracted_data,'visual')
pal_ = config.pal_
close_dates = config.close_dates

#feature for model
li_cont_select = config.li_cont_select
li_bin = config.li_bin
li_cat_select = config.li_cat_select

#create a director if not existing
if not os.path.exists(path_extracted_data_visual):
    os.makedirs(path_extracted_data_visual)
    
path_extracted_data_cleaning_model = os.path.join(path_extracted_data,'Cleaning','model')  

df_LF = pd.read_csv(os.path.join(path_extracted_data, id_run+'_LFCounterEqual0.csv'), parse_dates=['date'], dayfirst=True, sep=';') 
#x0 = df_daily.shape[0]
#fillnan with big number to it also means weird values if there is only nan
#df_daily['date_2remove_weirdLFcounter'] = df_daily.apply(lambda x: x['level'] in df_LF[(df_LF['LFCounter_nbr_equal0']>=10)&\
#(df_LF['sender']==x['TagID'])]['date'].tolist(), axis=1)
print(df_LF.shape)
display(df_LF.head(5))
display(df_LF.tail(5))
df_LF[df_LF['LFCounter_nbr_equal0']>0]

## Preprocess - add HenID,PenID

START_TIME = time.perf_counter()
p = glob.glob(os.path.join(path_initial_data, 'Barn 4 Pen*\TagUpdates\log*'))
#print('There is %d log files:\n  %s'%(len(p),'  \n  '.join(p)))
df = preprocessing_Origins(p, config, save=True, dodevice=True)
print(df.shape)
display(df.head(2))
display(df.tail(2))
END_TIME = time.perf_counter()
print ("Total running time: %.2f mn" %((END_TIME-START_TIME)/60))
#--> saves a datafarme named: _PreprocessRecords_forcleaning

# Model-based cleaning

In order to keep maximum information during video cleaning (i.e. each record), we kept consecutives equal zones, so the model was trained to say "correct record" even when the duration was small and the next zone was the same zone (instead of correct when the duration was longer and the next zone was a different one). 
Also, note that if we remove consecutives equal zone before the model filtering, then we might remove a correct record, while keeping a wrong record instead.
For these two reasons, we should remove consecutives equal zones AFTER model filtering and not during the general cleaning.
Note that the only difference with the preprocessing step in chapter 0, is that we add the milliseconds to the timsetamp in order to have the correct duration per stay first

#for being more reproductible, we open the file that was saved from cleaning
df = pd.read_csv(os.path.join(path_extracted_data, id_run+'_PreprocessRecords_forcleaning.csv'), sep=';', 
                 parse_dates=['Timestamp', 'date'], dayfirst=True) #index_col=0) 
#remove the entries that last less than a second
print(df.shape)
df = df.drop_duplicates(subset=['Timestamp','HenID'], keep='last')
print(min(df['date'].tolist()), max(df['date'].tolist()))
print(df.shape)
df.head(3)

#use same step as in cleaning-validation-effect paper
START_TIME = time.perf_counter()
df = cleaning_processing(df=df, date_min=dt.datetime(2020, 9, 29), date_max=dt.datetime(2021, 7, 26), config=config)
df['time'] = df['Timestamp'].map(lambda x: dt.datetime.time(x))
df.rename(columns={'Zone':'Trackingsystem_Zone'}, inplace=True) #thats how it was named in the model 
#add other variables
df['is_dur_bigger_'+str(nbr_maxdur2beremoved)] = df['duration'].map(lambda x: x>=nbr_maxdur2beremoved)
display(df['is_dur_bigger_'+str(nbr_maxdur2beremoved)].value_counts(normalize=True))
df['is_dur_smaller_60sec'] = df['duration'].map(lambda x: x<60)
display(df['is_dur_smaller_60sec'].value_counts(normalize=True))
print('saving...')
df.to_csv(os.path.join(path_extracted_data,'df_recordsmodelpreprocessing.csv'), sep=';', index=False)
END_TIME = time.perf_counter()
print ("Total running time: %.2f mn" %((END_TIME-START_TIME)/60))
print(df.shape)
df.head(3)

#download model
model = CatBoostClassifier()
path_ = r'G:\VPHI\Welfare\2- Research Projects\OFHE2.OriginsE2\DataOutput\TrackingSystem\CleaningAnalysis_\Cleaning\model'
model.load_model(os.path.join(path_,'FINAL_Catboost'));

#for more flexibility we open the file that was saved previously
#df = pd.read_csv(os.path.join(path_extracted_data, 'df_recordsmodelpreprocessing.csv'), sep=';', 
#                 parse_dates=['Timestamp','date'], dayfirst=True) 
df['is_day'] = df['Timestamp'].map(lambda x: is_day(x, config.dico_night_hour))
df['is_dur_smaller_60sec'] = df['duration'].map(lambda x: x<60)
display(df['is_dur_smaller_60sec'].value_counts())
df['duration_bounded_mn'] = df['duration'].map(lambda x: min(x,60*60)/60)
df['next_duration_bounded_mn'] = df['next_duration'].map(lambda x: min(x,60*60)/60)
df['previous_duration_bounded_mn'] = df['previous_duration'].map(lambda x: min(x,60*60)/60)
print(df.shape)
df.head(3)

df_ = df[(df['Trackingsystem_Zone']!='1_Zone')&\
         (df['HenID']=='hen_143')&(df['date'].isin([dt.datetime(2020,10,3), 
                                                    dt.datetime(2020,10,4)]))].copy()
print(df_.shape)
df_.head(3)

########### handle nan
#add info on nan values in the initial df, in order to replace at the end all prediction based on features involvig at least 
#one nan, by 1
#Indeed, for records with some nan features, we will predict them a 1 all the time its only 0,05% of the time , i.e. when its 
#first or last transition. Indeed, we dont want to loose any records!
df['nbr_nan'] = df.filter(li_cont_select+li_cat_select, axis=1).isnull().sum(axis=1).tolist()
display(df['nbr_nan'].value_counts())
#fillnan for simplicity, but we will remove those rows predictions anyway
#display BEFORE categoricalfeatures
display(df[li_cat_select].head(3))
df[li_cont_select] = df[li_cont_select].fillna(df[li_cont_select].mean().iloc[0])
#display AFTER categoricalfeatures
df[li_cat_select] = df[li_cat_select].fillna(df[li_cat_select].mode().iloc[0])
#show data
display(df[config.li_cont_select].head(3))
display(df[config.li_cat_select].head(3))

#Too slow with large dataset
START_TIME = time.perf_counter()
#predict, and if its wintergarten (i.e. 1_Zone) put =1
#to slow for large dataset, but more readable:
#df['model_prediction'] = df.apply(lambda x: model.predict(pd.DataFrame(x[li_cont_select+\
#                                                                         li_cat_select]).transpose())[0] if\
#                                  x['Trackingsystem_Zone']!='1_Zone' else 1, axis=1)
#predict, and if its wintergarten (i.e. 1_Zone) put =1
df_ = df[df['Trackingsystem_Zone']!='1_Zone'].copy()
predictions = model.predict(df_.filter(li_cont_select+li_cat_select, axis=1))
x0 = df[df['Trackingsystem_Zone']=='1_Zone'].shape[0]
df_['model_prediction'] = predictions
if df_[df_['model_prediction'].isnull()].shape[0]!=0:
    print('ERROR')
    sys.exit()
df = pd.merge(df, df_.filter(['model_prediction','HenID','Timestamp'], axis=1), on=['HenID','Timestamp'], how='left')
print(df[df['model_prediction'].isnull()].shape)
if df[df['model_prediction'].isnull()].shape[0]!=x0:
    print('ERROR: you have some nan in your predictions')
    sys.exit()
display(df['model_prediction'].value_counts(normalize=True))

#a bit more conservative than in the cleaning, as we have now longer durations, and we dont want to be able to add big errors

#predict false if <60seconds in WG or at night, else true
df.loc[(df['is_dur_smaller_60sec'])&((df['Trackingsystem_Zone']=='1_Zone')|(~df['is_day'])), 'model_prediction'] = 0
#always =1 if there was some nan values
display(df['model_prediction'].value_counts(normalize=True)) #in testing it was 4,9% of wrong records: 42/(42+821)*100)

#always =1 if >15mn or if there was some NAN values: correct if >15 as we saw its ~always the case, when not its because the  
#hen went in the zone that was wrongly registered and which was the latest registrations of that hen
df.loc[(df['is_dur_bigger_'+str(nbr_maxdur2beremoved)])|(df['nbr_nan']!=0), 'model_prediction'] = 1
display(df['model_prediction'].value_counts(normalize=True)) #in testing it was 4,9% of wrong records: 42/(42+821)*100)

#predict=1 if nan
df.loc[df['nbr_nan']!=0, 'model_prediction'] = 1
display(df['model_prediction'].value_counts(normalize=True)) #in testing it was 4,9% of wrong records: 42/(42+821)*100)

END_TIME = time.perf_counter()
print ("Total running time for prediction: %.2f mn" %((END_TIME-START_TIME)/60))
display(display(df.head(3)))
#prediction_proba = model.predict_proba(df[df['Trackingsystem_Zone']!='1_Zone'])
#df['model_prediction_proba0'] = [round(x[0],2) for x in prediction_proba]
#df['model_prediction_proba1'] = [round(x[1],2) for x in prediction_proba]

#will take TIME to save
#Note that we cant use the duration info anymore so we are filtering it
df.filter(['HenID','PenID','system','Trackingsystem_Zone','model_prediction','Timestamp','date'],
          axis=1).to_csv(os.path.join(path_extracted_data, id_run+'_Rawdata&Model.csv'), index=False, sep=';')

## Remove filtered records & consecutives equal zones

#for being more reproductible, we open the file that was saved from cleaning
df = pd.read_csv(os.path.join(path_extracted_data, id_run+'_Rawdata&Model.csv'), sep=';', parse_dates=['Timestamp','date'],
                 dayfirst=True) 
#remove predicted-wrong-records
print(df.shape)
print(df[df['model_prediction'].isnull()].shape)
df = df[df['model_prediction']!=0].copy()
print(df.shape)
#simply rename the tracking system zone into the model zone :) 
df.rename(columns={'Trackingsystem_Zone':'Zone'}, inplace=True)
print(df.shape)
df.head(3)

#######################################################################################################################
#ceiz = consecutives equal initial zone
print('-----------------------------------------------------------------------------------------------')
print('remove consecutives equal zone for the same hen.........')
print('-----------------------------------------------------------------------------------------------')

li_df = []
#more efficient to do it per hen, as it wont need to search in the whole dataframe, and we can simply shift the timestamp column
for i, df_hen in tqdm.tqdm(df.groupby(['HenID'])):
    #as the next record date (sort by date, then simply shift by one row and add nan at then end)
    df_hen = df_hen.sort_values(['Timestamp'], ascending=True)
    #same date, one must take the last recorded one & sorting by date might change it. Also it already should be sorted by date
    df_hen['next_record_date'] = df_hen['Timestamp'].tolist()[1:]+[np.nan]
    #compute duration
    df_hen['duration'] = df_hen.apply(lambda x: x['next_record_date']-x['Timestamp'], axis=1)
    #compute the last record date in order to put interzone also when the duration is >=nbr_sec_flickering1
    df_hen['previous_record_date'] = [np.nan]+df_hen['Timestamp'].tolist()[0:-1]
    #compute previous duration in order to put interzone also when the duration is >=nbr_sec_flickering1
    df_hen['previous_duration'] = [np.nan]+df_hen['duration'].tolist()[0:-1]
    #add next record for the impossible movement
    df_hen['next_zone'] = df_hen['Zone'].tolist()[1:]+[np.nan]
    #add previous record for the consecutives equal initial zones
    df_hen['previous_zone'] = [np.nan]+df_hen['Zone'].tolist()[0:-1]
    df_hen['previous_previous_zone'] = [np.nan]+df_hen['previous_zone'].tolist()[0:-1]
    li_df.append(df_hen)
#put again in one dataframe
df = pd.concat(li_df)
#dont care about the false positive warning   

#True if next zone is equal to the actual zone
df['correction_is_consecutive_equal_initial_zone'] = False
#if the previous zone is the same, then its cez that should be removed
df.loc[df['previous_zone']==df['Zone'], 'correction_is_consecutive_equal_initial_zone'] = True


################# save
df = df.sort_values(['Timestamp'], ascending=True)
#save to verify CEIZ
df.to_csv(os.path.join(path_extracted_data, id_run+'_records_GeneralCleaning_VERIFICATION_CEIZ.csv'), sep=';', index=False) 
df = df[~(df['correction_is_consecutive_equal_initial_zone'])]

#small verification
df_hen[['Timestamp','HenID','previous_zone','Zone']].head(10)

## Remove records in WG when WG was close

df['is_WG_open'] = df['Timestamp'].map(lambda x: is_WG_open(x, config, epsi_open=0, epsi_close=20))
#we keep the epsi_open=0 as the system appeared to always open between 20sec of 2mn AFTER the indicated time (from video 
#observation)
#we set epsi_close=20 as for video observation we saw that itwoul dbe between + and -15 mn than the indicated time. So we added
#some flexibility by puting 20mn.
print(df['is_WG_open'].value_counts(normalize=True))
display(df[(df['Zone']=='1_Zone')&(df['is_WG_open'])][['Timestamp','is_WG_open']].head(5))
display(df[(df['Zone']=='1_Zone')&(~df['is_WG_open'])][['Timestamp','is_WG_open']].head(5))

#remove WG when close
print(df.shape)
df = df[~((~df['is_WG_open'])&(df['Zone']=='1_Zone'))]
print(df.shape)

df = pd.read_csv(os.path.join(path_extracted_data, id_run+'_CLEANEDDATA.csv'), sep=';', parse_dates=['Timestamp', 'date'],
                dayfirst=True) 
print(df.shape)
df.head(3)

# remove transitions smaller than 15mn over night

########### remove transitions smaller than 15mn over night
#compute duration again as is not updated with alst cleaning
li_df = []
#more efficient to do it per hen, as it wont need to search in the whole dataframe, and we can simply shift the timestamp column
for i, df_hen in tqdm.tqdm(df.groupby(['HenID'])):
    #as the next record date (sort by date, then simply shift by one row and add nan at then end)
    df_hen = df_hen.sort_values(['Timestamp'], ascending=True)
    #same date, one must take the last recorded one & sorting by date might change it. Also it already should be sorted by date
    df_hen['next_record_date'] = df_hen['Timestamp'].tolist()[1:]+[np.nan]
    #compute duration
    df_hen['duration_mn'] = df_hen.apply(lambda x: (x['next_record_date']-x['Timestamp']).total_seconds()/60, axis=1)
    li_df.append(df_hen)
#put again in one dataframe
df = pd.concat(li_df)
df['is_day'] = df['Timestamp'].map(lambda x: is_day(x, config.dico_night_hour))
#display(df[~df['is_day']].head(3))
#df[~df['is_day']].tail(3)
print(df.shape)
display(df.head(3))
#df[(df['duration_mn']<15) & (df['is_day']==False)]
#df[(df['HenID']=='hen_1')&(df['Timestamp']<dt.datetime(2020,10,1,10,1,1))&(df['Timestamp']>dt.datetime(2020,9,30,17,50,1))]
df = df[~((df['duration_mn']<15) & (df['is_day']==False))]
print(df.shape)
df.head(3)

df.head(2)

# Save

print(df.shape)
df.head(3)
df.to_csv(os.path.join(path_extracted_data, id_run+'_CLEANEDDATA.csv'), sep=';', index=False) 

## MLP comparison (raw vs cleaned)

#for being more reproductible, we open the file that was saved from cleaning
df = pd.read_csv(os.path.join(path_extracted_data, id_run+'_Rawdata&Model.csv'), sep=';', parse_dates=['Timestamp','date'],
                 dayfirst=True) 
print(df.shape)
df.head(3)

df['model_prediction'].value_counts(normalize=True)

START_TIME = time.perf_counter()
#choose specific dates to plot, November is better as hens starts to move more and the algo have never seen it
dmin = dt.datetime(2021,5,10)
dmax = dt.datetime(2021,5,20)
#df_ = df[(df['date']<dt.datetime(2020,11,16))&(df['date']>dt.datetime(2020,11,10))]
df_ = df[(df['date']<=dmax)&(df['date']>=dmin)].copy()
print(df_.shape)
#only print the hens with at least one transition per day (simplicity and interestignly)
#to sort the yaxis
dico_zone_order = {'1_Zone':0, '2_Zone':1, '3_Zone':2, '4_Zone':3, '5_Zone':4}
li_ts = ['Trackingsystem_Zone','Model_Zone']
li_date = df_['date'].unique()
#for each batch draw 2 timeseries, the raw one and the model one
for henID, df_plt in tqdm.tqdm(df_.groupby(['HenID'])):
    df_plt = df_plt.sort_values(['Timestamp']).copy()
    c = len(li_ts) ; l = len(li_date)
    fig, ax = plt.subplots(figsize=(c*5, l*1))
    i = 1
    if len(df_plt['date'].unique())==len(li_date):
        mi = min(df_plt['Timestamp'].tolist()) ; ma = max(df_plt['Timestamp'].tolist())
        Daterange = pd.date_range(start = mi+dt.timedelta(seconds=(60-mi.second)), 
                                  end = ma-dt.timedelta(seconds=(ma.second+1)), 
                                  freq = 'S') 
        df_plt_ = df_plt.copy()
        df_plt_.set_index('Timestamp', inplace=True)
        df_plt_ = df_plt_.reindex(Daterange, method='ffill').reset_index()
        df_plt_.rename(columns={'index':'Timestamp'}, inplace=True)
        #add date again, as the reindexing also extended the date
        df_plt_['date'] = df_plt_['Timestamp'].map(lambda x: dt.datetime.date(x))
        #remove first & last date
        df_plt_ = df_plt_[~df_plt_['date'].isin([max(df_plt_['date'].tolist()),min(df_plt_['date'].tolist())])]
        #put xlabel into numbers for the ploting
        df_plt_['Trackingsystem_Zone'] = df_plt_['Trackingsystem_Zone'].map(lambda x: int(dico_zone_order[x]))       
        for d, df_plt__ in df_plt_.groupby(['date']):
            for v in li_ts:
                df_plt___ = df_plt__.copy()
                if v=='Model_Zone':
                    df_plt___ = df_plt___[df_plt___['model_prediction']==1]
                plt.subplot(l,c,i)
                plt.tight_layout(pad=0.3) #add spacing between each plot
                if i%2==1:
                    plt.yticks([0,1,2,3,4], ['Winter garden', 'Litter', 'Lower perch','Nestbox','Top floor'])
                else:
                    plt.yticks([0,1,2,3,4], ['','','','',''])
                if i in [1,2]:
                    plt.title(v.replace('_',' '), size=13)
                i = i+1
                plt.plot(df_plt___['Timestamp'].tolist(), df_plt___['Trackingsystem_Zone'].tolist(), 
                         color=pal_[v], linewidth=1)
                li_hour = pd.date_range(start = d,  end = d+dt.timedelta(days=1), freq = 'H')
                plt.xticks(li_hour , [str(i.hour)+'h' for i in li_hour], fontsize=5)                
                plt.yticks(fontsize=8)
                plt.xlabel(str(d).split('T')[0], size=7)
                plt.ylim(0, 4.2)  
        plt.savefig(os.path.join(path_extracted_data_visual, 
                                 henID+'_'+str(dmin).split(' ')[0]+'_'+str(dmax).split(' ')[0]+'.png'), 
                    format='png', bbox_inches='tight', dpi=300)
    plt.clf()
    plt.close("all")    
END_TIME = time.perf_counter()
print ("Total running time: %.2f mn" %((END_TIME-START_TIME)/60))
#hen4, 22.10.2020: flickering. solved?
